+++ packages/ui-datalist/src/modules/filters/modules/filterConfig/components/has-attachment/has-attachment-filter-value-preview.vue
@@ -3,22 +3,13 @@
 </template>
 
 <script lang="ts" setup>
-import { computed } from 'vue';
-import { useI18n } from 'vue-i18n';
-
-import { BooleanOptions } from '../../enums/options/BooleanFilterOptions';
+import { usePrettifyBooleanValuePreview } from '../_shared/composables/booleanFilterToolkit';
 
 const props = defineProps<{
   value: boolean;
 }>();
-const { t } = useI18n();
 
-const localeValue = computed(() => {
-  const match = BooleanOptions.find(
-    (el) => el.value === props.value.toString(),
-  );
-  return match ? t(match.locale) : '';
-});
+const { localeValue } = usePrettifyBooleanValuePreview(props.value);
 </script>
 
 <style scoped></style>